started working on differences for integer field
There is a lot of logic here.

I first check to see if they're different. if so, I go get the differences and then need to do an alter table query based on that.

I need to redo the alter table class because I have it altering the type AND other attributes as well. They must be done seperately.